Added ajax based autocomplete.
I may want to show some kind of spinny icon
while it's pulling info, but the latency isn't
too bad. Also, there are timeouts to give the
user time to type before we make a call.
We also check that after a series of keyups,
that the query has changed.