Added race condition filtering
There were times when I would press return, which
would fetch and briefly show the results, but then,
woops, the suggestions would show up and push the
info box down. Now info and sugs are setup to
be mutually exclusive. Also, searches will cancel
suggestions, and if the results for sugs comes back
after the search is already done, it will find that
the results are no longer needed and bail out.